requirements.txt
ag-ui-protocol>=0.1.0
openai>=1.0.0
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
typing-extensions>=4.5.0
python-dotenv>=1.0.0
requests>=2.31.0
sseclient-py>=1.7.2 